move logic of risky proposals to hooks
