+++ src/hooks/useRiskyProposals.ts
@@ -1,25 +1,11 @@
 import { useEffect, useState } from "react"
 import { RiskyProposal } from "constants/interfaces_v2"
-import { TraderPool, TraderPoolRiskyProposal } from "abi"
-import useContract from "hooks/useContract"
+import { useRiskyProposalContract } from "hooks/useContract"
 
 function useRiskyProposals(poolAddress?: string): RiskyProposal[] {
-  const [proposalAddress, setProposalAddress] = useState("")
   const [proposals, setProposals] = useState<RiskyProposal[]>([])
 
-  const traderPool = useContract(poolAddress, TraderPool)
-  const traderPoolRiskyProposal = useContract(
-    proposalAddress,
-    TraderPoolRiskyProposal
-  )
-
-  useEffect(() => {
-    if (!traderPool) return
-    ;(async () => {
-      const address = await traderPool.proposalPoolAddress()
-      setProposalAddress(address)
-    })()
-  }, [traderPool])
+  const [traderPoolRiskyProposal] = useRiskyProposalContract(poolAddress)
 
   useEffect(() => {
     if (!traderPoolRiskyProposal) return
@@ -36,13 +22,21 @@ export function useRiskyProposal(
   poolAddress?: string,
   index?: string
 ): RiskyProposal | undefined {
-  const proposals = useRiskyProposals(poolAddress)
+  const [proposal, setProposal] = useState<RiskyProposal | undefined>()
+  const [traderPoolRiskyProposal] = useRiskyProposalContract(poolAddress)
 
-  if (!index) {
-    return undefined
-  }
+  useEffect(() => {
+    if (!traderPoolRiskyProposal || !index) return
+    ;(async () => {
+      const data = await traderPoolRiskyProposal.getProposalInfos(
+        index,
+        parseFloat(index) + 1
+      )
+      setProposal(data[0])
+    })()
+  }, [index, traderPoolRiskyProposal])
 
-  return proposals[index]
+  return proposal
 }
 
 export default useRiskyProposals
