+++ src/pages/CreateRiskyProposal/index.tsx
@@ -1,8 +1,6 @@
-import { FC, useEffect, useState } from "react"
+import { FC, useState } from "react"
 import { Flex } from "theme"
-import { BigNumber, ethers } from "ethers"
-import getTime from "date-fns/getTime"
-import { addDays, format } from "date-fns/esm"
+import { format } from "date-fns/esm"
 import {
   Navigate,
   Route,
@@ -11,8 +9,6 @@ import {
   useParams,
 } from "react-router-dom"
 import { useSelector } from "react-redux"
-import { useWeb3React } from "@web3-react/core"
-import { TraderPoolRiskyProposal, BasicTraderPool } from "abi"
 import { createClient, Provider as GraphProvider } from "urql"
 
 import Header from "components/Header/Layout"
@@ -28,31 +24,20 @@ import DatePicker from "components/DatePicker"
 import Payload from "components/Payload"
 import TransactionError from "modals/TransactionError"
 
-import { usePoolContract, usePoolQuery, useTraderPool } from "hooks/usePool"
+import { usePoolContract } from "hooks/usePool"
 import { Token } from "constants/interfaces"
 import { selectWhitelist } from "state/pricefeed/selectors"
-import useContract, {
-  useBasicPoolContract,
-  useERC20,
-  useRiskyProposalContract,
-} from "hooks/useContract"
+import { useERC20 } from "hooks/useContract"
 import useTokenPriceOutUSD from "hooks/useTokenPriceOutUSD"
 
-import {
-  expandTimestamp,
-  formatBigNumber,
-  normalizeBigNumber,
-  shortTimestamp,
-  parseTransactionError,
-} from "utils"
-
-import getReceipt from "utils/getReceipt"
+import { expandTimestamp, formatBigNumber, normalizeBigNumber } from "utils"
 
 import back from "assets/icons/angle-left.svg"
 import close from "assets/icons/close-big.svg"
 import calendar from "assets/icons/calendar.svg"
 
 import faqText from "./faq"
+import useCreateRiskyProposal from "./useCreateRiskyProposal"
 
 import {
   Container,
@@ -76,7 +61,6 @@ import {
   Label,
   White,
   Grey,
-  contentVariants,
 } from "./styled"
 
 const poolsClient = createClient({
@@ -84,167 +68,6 @@ const poolsClient = createClient({
 })
 
 const isFaqRead = localStorage.getItem("risky-proposal-faq-read") === "true"
-
-const useCreateRiskyProposal = (
-  poolAddress?: string,
-  tokenAddress?: string
-): [
-  {
-    error: string
-    isSubmiting: boolean
-    baseTokenPrice?: BigNumber
-    lpAvailable?: BigNumber
-    lpAmount: string
-    timestampLimit: number
-    investLPLimit: string
-    maxTokenPriceLimit: string
-    instantTradePercentage: number
-  },
-  {
-    setSubmiting: (value: boolean) => void
-    setError: (value: string) => void
-    setLpAmount: (value: string) => void
-    setTimestampLimit: (timestamp: number) => void
-    setInvestLPLimit: (value: string) => void
-    setMaxTokenPriceLimit: (value: string) => void
-    setInstantTradePercentage: (percent: number) => void
-    handleSubmit: () => void
-  }
-] => {
-  const { account, library } = useWeb3React()
-  const initialTimeLimit = shortTimestamp(getTime(addDays(new Date(), 30)))
-  const [riskyProposal] = useRiskyProposalContract(poolAddress)
-
-  const basicTraderPool = useBasicPoolContract(poolAddress)
-  const traderPool = useTraderPool(poolAddress)
-
-  const [error, setError] = useState("")
-  const [isSubmiting, setSubmiting] = useState(false)
-  const [lpAmount, setLpAmount] = useState("")
-  const [timestampLimit, setTimestampLimit] = useState(initialTimeLimit)
-  const [investLPLimit, setInvestLPLimit] = useState("")
-  const [maxTokenPriceLimit, setMaxTokenPriceLimit] = useState("")
-  const [instantTradePercentage, setInstantTradePercentage] = useState(0)
-  const [baseTokenPrice, setBaseTokenPrice] = useState<BigNumber | undefined>()
-  const [lpAvailable, setLpAvailable] = useState<BigNumber | undefined>()
-
-  useEffect(() => {
-    if (!riskyProposal || !tokenAddress || tokenAddress.length !== 42) return
-
-    const getCreatingTokensInfo = async () => {
-      const tokens = await riskyProposal.getCreationTokens(
-        tokenAddress,
-        ethers.utils.parseEther("1").toHexString(),
-        ethers.utils.parseUnits("100", 27).toHexString(),
-        []
-      )
-      setBaseTokenPrice(tokens.positionTokenPrice)
-    }
-
-    getCreatingTokensInfo().catch(console.error)
-  }, [riskyProposal, tokenAddress])
-
-  // get risky pool address
-  // useEffect(() => {
-  //   if (!basicTraderPool) return
-
-  //   const getProposalPoolAddress = async () => {
-  //     const address = await basicTraderPool.proposalPoolAddress()
-  //     setRiskyProposalAddress(address)
-  //   }
-
-  //   getProposalPoolAddress().catch(console.error)
-  // }, [basicTraderPool])
-
-  useEffect(() => {
-    if (!traderPool || !account) return
-
-    const getBalance = async () => {
-      const lpAvailable: BigNumber = await traderPool.balanceOf(account)
-      setLpAvailable(lpAvailable)
-    }
-
-    getBalance().catch(console.error)
-  }, [traderPool, account])
-
-  const handleSubmit = () => {
-    if (!basicTraderPool || !traderPool || !riskyProposal) return
-
-    const createRiskyProposal = async () => {
-      setSubmiting(true)
-      setError("")
-      const amount = ethers.utils.parseEther(lpAmount).toHexString()
-      const percentage = ethers.utils
-        .parseUnits(instantTradePercentage.toString(), 25)
-        .toHexString()
-
-      const divests = await traderPool.getDivestAmountsAndCommissions(
-        account,
-        amount
-      )
-
-      const tokens = await riskyProposal.getCreationTokens(
-        tokenAddress,
-        divests.receptions.baseAmount,
-        percentage,
-        []
-      )
-
-      const createReceipt = await basicTraderPool.createProposal(
-        tokenAddress,
-        amount,
-        [
-          timestampLimit,
-          ethers.utils.parseUnits(investLPLimit, 18).toHexString(),
-          ethers.utils.parseUnits(maxTokenPriceLimit, 18).toHexString(),
-        ],
-        percentage,
-        divests.receptions.receivedAmounts,
-        tokens[0],
-        []
-      )
-
-      await getReceipt(library, createReceipt.hash)
-      setSubmiting(false)
-    }
-
-    createRiskyProposal().catch((error) => {
-      setSubmiting(false)
-      console.error(error)
-
-      if (!!error && !!error.data && !!error.data.message) {
-        setError(error.data.message)
-      } else {
-        const errorMessage = parseTransactionError(error.toString())
-        !!errorMessage && setError(errorMessage)
-      }
-    })
-  }
-
-  return [
-    {
-      lpAmount,
-      error,
-      isSubmiting,
-      lpAvailable,
-      baseTokenPrice,
-      timestampLimit,
-      investLPLimit,
-      maxTokenPriceLimit,
-      instantTradePercentage,
-    },
-    {
-      setError,
-      setSubmiting,
-      setLpAmount,
-      setTimestampLimit,
-      setInvestLPLimit,
-      setMaxTokenPriceLimit,
-      setInstantTradePercentage,
-      handleSubmit,
-    },
-  ]
-}
 
 const CreateRiskyProposal: FC = () => {
   const { tokenAddress, poolAddress } = useParams()
@@ -520,20 +343,8 @@ const CreateRiskyProposal: FC = () => {
                   />
                 ) : (
                   <>
-                    <CardHeader
-                      initial="hidden"
-                      animate="visible"
-                      variants={contentVariants}
-                    >
-                      {stepComponents["1"].header}
-                    </CardHeader>
-                    <Content
-                      initial="hidden"
-                      animate="visible"
-                      variants={contentVariants}
-                    >
-                      {stepComponents["1"].content}
-                    </Content>
+                    <CardHeader>{stepComponents["1"].header}</CardHeader>
+                    <Content>{stepComponents["1"].content}</Content>
                   </>
                 )
               }
