+++ src/state/pools/hooks.tsx
@@ -27,8 +27,6 @@ import {
   BasicPositionsQuery,
   getPoolsQueryVariables,
 } from "queries"
-import { BigNumber, ethers, FixedNumber } from "ethers"
-import { usePoolContract, useTraderPool } from "hooks/usePool"
 
 /**
  * Returns top members filter state variables and setter
@@ -96,31 +94,6 @@ export function usePriceHistory(
   }, [pool])
 
   return history
-}
-
-export function usePoolPrice(address: string | undefined) {
-  const traderPool = useTraderPool(address)
-  const [priceUSD, setPriceUSD] = useState(ethers.utils.parseEther("1"))
-  const [priceBase, setPriceBase] = useState(ethers.utils.parseEther("1"))
-
-  useEffect(() => {
-    if (!traderPool) return
-    ;(async () => {
-      const poolInfo = await traderPool.getPoolInfo()
-      if (poolInfo.lpSupply.gt("0")) {
-        const base = FixedNumber.fromValue(poolInfo.totalPoolBase, 18)
-        const usd = FixedNumber.fromValue(poolInfo.totalPoolUSD, 18)
-        const supply = FixedNumber.fromValue(poolInfo.lpSupply, 18)
-
-        const usdPrice = usd.divUnsafe(supply)
-        const basePrice = base.divUnsafe(supply)
-        setPriceUSD(BigNumber.from(usdPrice._hex))
-        setPriceBase(BigNumber.from(basePrice._hex))
-      }
-    })()
-  }, [traderPool])
-
-  return { priceUSD, priceBase }
 }
 
 // Hook that handles fetching and storing pools
